+++ DBClientFiles.NET/Parsing/File/WDBC/Header.cs
@@ -32,7 +32,7 @@ namespace DBClientFiles.NET.Parsing.File.WDBC
         public int MinIndex => throw new NotImplementedException();
         public int MaxIndex => throw new NotImplementedException();
         public int CopyTableLength => throw new NotImplementedException();
-        public int IndexColumn => throw new NotImplementedException();
+        public short IndexColumn => throw new NotImplementedException();
         public bool HasIndexTable => throw new NotImplementedException();
         public bool HasForeignIds => throw new NotImplementedException();
         public bool HasOffsetMap => throw new NotImplementedException();
